This should work (and work when packaged)
Removed the native directory. Made the tests run expecting the same
structure as what we distribute. That same structure is created by
either running `yarn compile` or `node-pre-gyp install` (the second
being how and end user gets the binary).

The publish structure is the same other than the new entry point, so
that should continue working.